requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_text_splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
scikit-learn
ipykernel
numpy
pandas
duckdb
numexpr
pytube
duckduckgo-search
unstructured
mysql-connector-python
SQLAlchemy
